+++ History.md
@@ -1,4 +1,10 @@
 
+0.2.0 / 2013-12-08 
+==================
+
+ * Rebuild captions correctly on every style change. Fixes BRBPL-3923.
+ * Restructure code so it's more maintainable.  - Breaking monolithic captionator.js into separate files.
+
 0.1.2 / 2013-12-06 
 ==================
  * further support for updating closed captions styles.